Add structure to binary operations.
All binary operations are left-associative, at least for now.

Note, this involved making the rules non-self recursive.  They consume
arbitrary amounts of repetitions of their operators using the greedy
"many-times" operator, "*", so the self-recursion was redundant.  (It
was also causing right-associativity, which is not what we usually
want.)

Note, these are still recursive in the sense that any expression can
be surrounded with parentheses, and then any operation can be applied
to that atomic unit.  We no valid Decaf programs can fail due to this
change.