Added Subexpression.java to optimization package
This will act as a unique identifier for NativeExpressions. It
wraps around the Native Expression and the Scope. We need the
scope to distinguish between cases like the following:

...
if (...) {
   int b;
   ...
}
if (...) {
   int b;
}

Also includes a getMostGeneralScope function, which will be useful
in the following case:

int a,b,c,d;
...
a = c+d
if (...) {
    b = c+d
}

Should consider calling getSupremumScope instead of
getMostGeneralScope.