Added Availability.java
Disclaimer: I'm not sure this is the best way to go about it, but
I wanted to start writing up some of the patterns necessary for
this kind of analysis.

This class performs the Lattice calculations for Available
Expressions. Instead of using bit vectors, we are using sets of
expressions. This should make the code itself easier to write,
since we figure bit vectors are an abstraction that simply makes
the concept easier to teach.