Represent method-level Control Flow Graphs with `FlowGraph<Instruction>`s.
Previously, every CFGs was represented by a `BiTerminalGraph`.  But,
BiTermminalGraphs depend on `ControlFlowNode`s, and we are working to
eliminate `ControlFlowNode`s.

Now, every method-level CFG is represented by a
`FlowGraph<Instruction>`. All classes that work with method-level
CFGs, namely `MethodBlockPrinter` have been re-written to work with
the new representation.

For now, `BiTerminalGraphs` are still used to construct the
method-level CFG.  Eventually, all uses of `BiTerminalGraph` will be
replaced with the new representaiton.  But, for the interim period,
`BiTerminalGraph` has been given the `asFlowGraph` method.

As part of this change,`MethodBlockPrinter` has been given a big
overhaul, which includes the elimination of the `VisitedSet` and
`SourceCounter` classes.

Note that this change breaks the CFG-printing functionality, for now.