Created AnalysisSpec class.
Modified AvailabilityCalculator to be a general forward
propagator. In its constructor, it now needs to be passed an
AnalysisSpec, which specifies certain functions necessary for
availability calculations.

This will allow us to convert the AvailabilityCalculator into a
general propagator class from which we can perform our fixed point
algorithms for reaching definition, available expressions, and
liveness.