Update CommonExpressionEliminator to use new scopes.
Previously, the CEE mutated scopes... which went along nicely with its
mutation of DFGs!  But, now that we are no longer mutating
DFGs, we should no longer mutate Scopes.

This commit causes a new, augmented Scope tree to be generated, and for
the new DFG to use these new scopes instead of the old ones.