convert ScopedLocation to ScopedVariable.
This allows us to avoid the aliasing issues that arise when we treat
each array slot as an individual unit.  (For example, if we try to do
Constant Propagation on `A[5] = 10; i = 5; A[i] = 20; x = A[5];`, we
would accidentally replace the last assignment with `x = 10`, because
we would not realize that `A[5]` is now `20`.

Now, we treat the array `A` as a single unit.