Dominator trees implemented
There is now a new target for printing out dominator trees: "dom".
Uses mostly the machinery that we already had in place. The main
new code is determining which edges is the dominator tree are
worth printing.

They can be generated using "-t dom". Will be useful for detecting
loops.

Needs further testing for correctness.