Fixed Dominator Trees.
The problem was that nodes that are header of loops need to
disregard sources coming from back edges.

A problem for the printer was that nodes the way we determined if
an edge should be on the dominator tree was to look at the
children of the Dominator D and see if any of those nodes also
dominate the Conquered C. This would lead to problems with
branching nodes.

See ASCII illustration for Dominator D and Conquered C:
D - B - A - C
 \- E -/
Tree would print D-C edge, despite there being a D-A edge.

The solution is to also look at the parents of C and not draw the
edge if only one of the parents of C is dominated.