Fixed the Human-Readable Dominator Tree Generator.
It didn't take into account if a dominator was an immediate
dominator. Now it immediately registers that it should be shown.

Moved it to its own class. DominatorSpec doesn't need to know how
a tree is printed.

As an aside, this added complexity would've been necessary to
address in the normal running of the algorithm also, except
instead of at the parents-of-the-conquered level, at the
children-of-the-dominator level.